+++ database/migrations/2022_11_12_151022_SaranaTable.php
@@ -16,7 +16,7 @@
 				$table->string('keterangan',200)->nullable();
 				$table->dateTime('update_terakhir')->nullable();
 				$table->string('update_oleh',32)->nullable();
-				$table->integer('id_sarana')->primary();
+				$table->bigIncrements('id_sarana')->primary();
 			});
 		}
