+++ database/migrations/2022_11_12_151030_StatusSaranaTable.php
@@ -22,7 +22,7 @@
 				$table->string('keterangan',200)->nullable();
 				$table->dateTime('update_terakhir')->nullable();
 				$table->string('update_oleh',32)->nullable();
-				$table->integer('id_status_sarana')->primary();
+				$table->bigIncrements('id_status_sarana')->primary();
 			});
 		}
