+++ database/migrations/2022_11_12_151073_StatusPrasaranaRuangTable.php
@@ -27,7 +27,7 @@
 				$table->string('keterangan',200)->nullable();
 				$table->dateTime('update_terakhir')->nullable();
 				$table->string('update_oleh',32)->nullable();
-				$table->integer('id_status_prasarana_ruang')->primary();
+				$table->bigIncrements('id_status_prasarana_ruang')->primary();
 			});
 		}
