+++ database/migrations/2022_11_12_151092_InstrukturTable.php
@@ -21,7 +21,7 @@
 				$table->string('keterangan',200)->nullable();
 				$table->dateTime('update_terakhir')->nullable();
 				$table->string('update_oleh',32)->nullable();
-				$table->integer('id_instruktur')->primary();
+				$table->bigIncrements('id_instruktur')->primary();
 			});
 		}
