+++ database/migrations/2022_11_12_151139_PesertaPembinaanKepribadianTable.php
@@ -18,7 +18,7 @@
 				$table->string('predikat',50)->nullable();
 				$table->dateTime('update_terakhir')->nullable();
 				$table->string('update_oleh',32)->nullable();
-				$table->integer('id_peserta_pk')->primary();
+				$table->bigIncrements('id_peserta_pk')->primary();
 			});
 		}
